+++ kurtosis-module/impl/module_io/default_params.go
@@ -15,7 +15,7 @@ func GetDefaultExecuteParams() *ExecuteParams {
 			SlotsPerEpoch:                      32,
 			AltairForkEpoch:                    1,
 			MergeForkEpoch:                     2,
-			TotalTerminalDifficulty:            60000000,
+			TotalTerminalDifficulty:            100000000,
 			NumValidatorKeysPerNode:            64,
 			PreregisteredValidatorKeysMnemonic: "giant issue aisle success illegal bike spike question tent bar rely arctic volcano long crawl hungry vocal artwork sniff fantasy very lucky have athlete",
 		},