+++ kurtosis-module/impl/participant_network/cl/lighthouse/lighthouse_launcher.go
@@ -94,6 +94,7 @@ func (launcher *LighthouseCLClientLauncher) Launch(
 		return nil, stacktrace.NewError("Expected new Lighthouse Beacon service to have port with ID '%v', but none was found", beaconHttpPortID)
 	}
 
+	// TODO This will return a 503 when genesis isn't yet ready!!! Need to fix this somehow
 	beaconRestClient := cl_client_rest_client.NewCLClientRESTClient(beaconServiceCtx.GetPrivateIPAddress(), beaconHttpPort.GetNumber())
 	if err := availability_waiter.WaitForBeaconClientAvailability(beaconRestClient, maxNumHealthcheckRetries, timeBetweenHealthcheckRetries); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred waiting for the new Lighthouse Beacon node to become available")