+++ kurtosis-module/impl/participant_network/el/nethermind/nethermind_mining_waiter.go
@@ -6,6 +6,7 @@ type nethermindMiningWaiter struct {}
 
 // TODO Nethermind actually doesn't do any mining!!!! How do we handle this???
 func (n nethermindMiningWaiter) WaitForMining(numRetries uint32, timeBetweenRetries time.Duration) error {
+	// TODO use the same Geth API to verify that blockNumber is increasing
 	return nil
 }